Convert token identifiers to Long for torch < 1.8.0
Since PyTorch 1.8.0 token identifiers can be both Int and Long for
embedding lookups. Prior versions only support Long. Since we still
support older versions, convert token identifiers to Long for
compatibility.

This fixes the incompatibility with older versions introduced in #289.

+++ spacy_transformers/layers/transformer_model.py
@@ -205,7 +205,8 @@ def _convert_transformer_inputs(model, wps: WordpieceBatch, is_train):
 
     hf_device = model.shims[0]._hfmodel.transformer.device
     kwargs = {
-        "input_ids": xp2torch(wps.input_ids).to(device=hf_device),
+        # Note: remove conversion to long when PyTorch >= 1.8.0.
+        "input_ids": xp2torch(wps.input_ids).long().to(device=hf_device),
         "attention_mask": xp2torch(wps.attention_mask).to(device=hf_device),
     }
     if wps.token_type_ids is not None:
